# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FieldForge API
  version: 0.3.0
  description: |
    Multitenant ERP API for field services (US).
    Protected routes require `Authorization: Bearer <token>` and `X-Tenant-ID` header.

    ## Cross-cutting middleware

    - **Rate limiting:** Public routes (`/auth/signup`, `/auth/login`) are limited per client IP.
      Authenticated routes are limited per tenant. Responses include `X-RateLimit-Limit`,
      `X-RateLimit-Remaining`, and `X-RateLimit-Reset`. HTTP 429 includes `Retry-After`.
    - **Idempotency:** `POST` requests on protected routes accept optional `Idempotency-Key`.
      Replays within 24h return the cached response with `Idempotency-Replayed: true`.
    - **Domain events:** Mutations publish to the transactional outbox (`domain_events`);
      a background worker marks events `processed` after dispatch.
servers:
  - url: https://api.fieldforge.com/api/v1
    description: Production
  - url: http://localhost:8081/api/v1
    description: Local API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PlatformConfig:
      type: object
      properties:
        brand_overrides: { type: object, additionalProperties: true }
        feature_flags:
          type: object
          additionalProperties: { type: boolean }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /auth/login:
    post:
      summary: Authenticate user
      tags: [Auth]
      description: Rate limited per client IP (default 20 req/min).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: JWT token
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/me:
    get:
      summary: Current user profile
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantID" }]
      responses:
        "200":
          description: User profile
  /users:
    get:
      summary: List workspace users for the current tenant
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantID" }]
      responses:
        "200":
          description: Tenant-scoped user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        email: { type: string, format: email }
                        role: { type: string }
                        first_name: { type: string }
                        last_name: { type: string }
                        created_at: { type: string, format: date-time }
...
  /public/plans:
    get:
      summary: Active SaaS plans for signup and marketing (no auth)
      tags: [Platform]
      description: Returns active platform plans. Falls back to config pricing when DB catalog is empty.
      responses:
        "200":
          description: Public plan catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PublicPlan" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /platform/auth/login:
    post:
      summary: Authenticate platform operator
      tags: [Platform]
      description: Rate limited per client IP. Issues a platform JWT without tenant scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Platform JWT and admin profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  admin:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      email: { type: string, format: email }
        "401":
          description: Invalid credentials
        "429":
          $ref: "#/components/responses/RateLimited"
  /platform/plans:
    get:
      summary: List all platform plans (admin)
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Plan catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PlatformPlan" }
        "401":
          description: Unauthorized
        "403":
          description: Not a platform admin
    post:
      summary: Create platform plan
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name]
              properties:
                id: { type: string }
                name: { type: string }
                description: { type: string }
                price_monthly_cents: { type: integer, nullable: true }
                stripe_price_id: { type: string }
                features: {}
                active: { type: boolean }
                sort_order: { type: integer }
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformPlan" }
        "400":
          description: Validation error
        "403":
          description: Not a platform admin
  /platform/plans/{id}:
    patch:
      summary: Update platform plan
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                price_monthly_cents: { type: integer, nullable: true }
                stripe_price_id: { type: string }
                features: {}
                active: { type: boolean }
                sort_order: { type: integer }
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformPlan" }
        "404":
          description: Plan not found
        "403":
          description: Not a platform admin
    delete:
      summary: Delete platform plan
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Plan deleted
        "404":
          description: Plan not found
        "403":
          description: Not a platform admin
  /platform/promotions:
    get:
      summary: List promotions
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Promotion list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PlatformPromotion" }
    post:
      summary: Create promotion
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                description: { type: string }
                discount_percent: { type: integer, nullable: true }
                discount_cents: { type: integer, nullable: true }
                plan_id: { type: string }
                starts_at: { type: string, format: date-time, nullable: true }
                ends_at: { type: string, format: date-time, nullable: true }
                max_redemptions: { type: integer, nullable: true }
                active: { type: boolean }
      responses:
        "201":
          description: Promotion created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformPromotion" }
  /platform/promotions/{id}:
    patch:
      summary: Update promotion
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                description: { type: string }
                discount_percent: { type: integer, nullable: true }
                discount_cents: { type: integer, nullable: true }
                plan_id: { type: string }
                starts_at: { type: string, format: date-time, nullable: true }
                ends_at: { type: string, format: date-time, nullable: true }
                max_redemptions: { type: integer, nullable: true }
                active: { type: boolean }
      responses:
        "200":
          description: Promotion updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformPromotion" }
        "404":
          description: Promotion not found
    delete:
      summary: Delete promotion
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Promotion deleted
        "404":
          description: Promotion not found
  /platform/landing-content:
    get:
      summary: List landing page content blocks
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: section
          in: query
          required: false
          schema:
            type: string
            enum: [hero, features, pricing, footer]
      responses:
        "200":
          description: Landing content blocks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/LandingContentBlock" }
    post:
      summary: Create landing content block
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [section]
              properties:
                section:
                  type: string
                  enum: [hero, features, pricing, footer]
                block_key: { type: string }
                content: { type: object, additionalProperties: true }
                active: { type: boolean }
                sort_order: { type: integer }
      responses:
        "201":
          description: Block created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LandingContentBlock" }
  /platform/landing-content/{id}:
    patch:
      summary: Update landing content block
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                section:
                  type: string
                  enum: [hero, features, pricing, footer]
                block_key: { type: string }
                content: { type: object, additionalProperties: true }
                active: { type: boolean }
                sort_order: { type: integer }
      responses:
        "200":
          description: Block updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LandingContentBlock" }
        "404":
          description: Block not found
    delete:
      summary: Delete landing content block
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Block deleted
        "404":
          description: Block not found
  /platform/config:
    get:
      summary: Get platform config overrides
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Platform config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformConfig" }
    patch:
      summary: Update platform config overrides
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brand_overrides: { type: object, additionalProperties: true }
                feature_flags:
                  type: object
                  additionalProperties: { type: boolean }
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformConfig" }
...
  /platform/tenants:
    get:
      summary: List all tenants (registry)
      tags: [Platform]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Tenant registry
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PlatformTenant" }
...